Split out omitUrlParams tests.
